Add integration test for ProductService and UserService

+++ src/Service/UserService.php
@@ -80,10 +80,10 @@ class UserService extends BaseService
         // If parent is admin he can create user and client
         if(in_array('ROLE_BILEMO', $parent->getRoles())) {
             $role[] = 'ROLE_USER';
-            if(null === $roles) {
+            if(null !== $roles) {
                 //client have no parent
                 $parent = null;
-                $role[] = 'ROLE_CLIENT';
+                $role = $roles;
             }
         }
 
@@ -103,9 +103,9 @@ class UserService extends BaseService
     public function editUser(User $user, array $data, User $parent = null): User
     {
 
-        /*if(isset($parent) && $user->getClient() != $parent) {
-            return new \Exception('User and client not match', Response::HTTP_BAD_REQUEST);
-        }*/
+        if(isset($parent) && $user->getClient() !== $parent) {
+            throw new \Exception('User and client not match', Response::HTTP_BAD_REQUEST);
+        }
 
         $this->addUserInfos($user, $data);
 
@@ -121,9 +121,9 @@ class UserService extends BaseService
     public function deleteUser(User $user, User $parent = null)
     {
 
-        /*if(isset($parent) && $user->getClient() !== $parent) {
-            return new \Exception('User and client not match', Response::HTTP_BAD_REQUEST);
-        }*/
+        if(isset($parent) && $user->getClient() !== $parent) {
+            throw new \Exception('User and client not match', Response::HTTP_BAD_REQUEST);
+        }
 
         $entityManager = $this->managerRegistry->getManager();
         $entityManager->remove($user);
